+++ src/controller/index.ts
@@ -3,7 +3,7 @@ export { theme }
 export * from "./auth"
 export * from "./firebase"
 export * from "./config"
-export * from "./note"
-export * from "./note-hooks"
-export * from "./collection-hook"
-export * from "./collection"
+export * from "./edit/note"
+export * from "./edit/note-hook"
+export * from "./edit/collection"
+export * from "./edit/collection-hook"
